+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 - Display the total distance on each marker (closes #2).
 - Display altitude in marker's text (closes #4).
 - Improve the marker bubble design.
+- Adjust the code to fit into PythonAnywhere hosting.
 - [dev] Upgrade JS dependency `leaflet` from `1.7.1` to `1.9.4`.
 - [dev] Upgrade JS dependency `leaflet-routing-machine` from `3.2.0` to `3.2.12`.
 - [dev] Upgrade Python dependency `bottle` from `0.12.19` to `0.12.25`.
+- [dev] Add Python dependency `paste` `3.5.3` to serve the local HTTP server.
 - [dev] Introduce developement dependencies to lint the code.
 - [dev] Ignore trace details (keep only a final ZIP'ed file containing all traces).